\name{house}

\docType{data}

\alias{house}

\title{House dataset}

\description{The House dataset.}

\usage{data(house)}

\format{Dataframe with 26 observations and 14 variables.}

\details{

  Variables in order: price, bdr, flr, fp, rms, st, lot, tax, bth, con,
  gar, cdn, l1, l2.

}

\source{\url{http://utstat.toronto.edu/~srivasta/sasreg/CH3E1.SAS}}

\references{
  Sen, A. and Srivastava, M. (1990). Regression Analysis. Springer.
}

\keyword{datasets}
